+++ src/js/core/mediaelement.js
@@ -308,7 +308,7 @@ class MediaElement {
 					const
 						capName = `${propName.substring(0, 1).toUpperCase()}${propName.substring(1)}`,
 						getFn = () => (t.mediaElement.renderer !== undefined && t.mediaElement.renderer !== null &&
-						typeof t.mediaElement.renderer[`get${capName}`] === 'function') ? t.mediaElement.renderer[`get${capName}`]() : null,
+							typeof t.mediaElement.renderer[`get${capName}`] === 'function') ? t.mediaElement.renderer[`get${capName}`]() : null,
 						setFn = (value) => {
 							if (t.mediaElement.renderer !== undefined && t.mediaElement.renderer !== null &&
 								typeof t.mediaElement.renderer[`set${capName}`] === 'function') {
@@ -389,30 +389,9 @@ class MediaElement {
 			},
 			triggerAction = (methodName, args) => {
 				try {
-					const response = t.mediaElement.renderer[methodName](args);
-					if (response && typeof response.then === 'function') {
-						response.catch((e) => {
-							// Sometimes, playing media might throw `DOMException: The play() request was interrupted`.
-							// If so, pause and re-execute the action.
-							if (methodName === 'play') {
-								if (t.mediaElement.paused) {
-									setTimeout(() => {
-										const tmpResponse = t.mediaElement.renderer.play();
-										if (tmpResponse !== undefined) {
-											// Final attempt: just pause the media if it's not paused
-											tmpResponse.catch(() => {
-												if (!t.mediaElement.renderer.paused) {
-													t.mediaElement.renderer.pause();
-												}
-											});
-										}
-									}, 150);
-								}
-							} else {
-								return t.mediaElement.generateError(e, mediaFiles);
-							}
-						});
-					}
+					setTimeout(() => {
+						t.mediaElement.renderer[methodName](args);
+					}, (methodName === 'play' ? 150 : 0));
 				} catch (e) {
 					t.mediaElement.generateError(e, mediaFiles);
 				}