+++ changelog.md
@@ -16,6 +16,7 @@
 * Added `forceLive` configuration to hide progress bar and display `Live Broadcast` even when `duration` is a valid number @rafa8626
 * Fixed issues with Flash HLS renderer related to restarting video once ended @rafa8626
 * Added missing translations @rafa8626
+* Fixed issue with native methods assigned to MediaElement shim related to `The play() request was interrupted` error @rafa8626
 
 *4.2.3 (2017/07/22)*